+++ src/lang/deprecated.js
@@ -111,3 +111,34 @@ export class GUI_Object extends UISpriteElement {
         super(x, y, settings);
     }
 }
+
+
+/**
+ * return the width of the system Canvas
+ * @public
+ * @name getWidth
+ * @class
+ * @memberof Renderer#
+ * @deprecated since 15.12.0
+ * @see width
+ */
+Renderer.prototype.getWidth = function()  {
+    warning("getWidth", "width", "15.12.0");
+    return this.width;
+};
+
+/**
+ * return the height of the system Canvas
+ * @public
+ * @name getHeight
+ * @class
+ * @memberof Renderer#
+ * @deprecated since 15.12.0
+ * @see height
+ */
+Renderer.prototype.getHeight = function()  {
+    warning("getHeight", "height", "15.12.0");
+    return this.height;
+};
+
+