+++ src/state/stage.js
@@ -95,8 +95,8 @@ export default class Stage {
         // empty or no default camera
         if (this.cameras.has("default") === false) {
             if (typeof default_camera === "undefined") {
-                let width = renderer.getWidth();
-                let height = renderer.getHeight();
+                let width = renderer.width;
+                let height = renderer.height;
                 // new default camera instance
                 default_camera = new Camera2d(0, 0, width, height);
             }