+++ src/video/renderer.js
@@ -107,6 +107,30 @@ export default class Renderer {
     }
 
     /**
+     * return the height of the canvas which this renderer draws to
+     * @returns {number} height of the system Canvas
+     */
+    get height() {
+        return this.getCanvas().height;
+    }
+
+    set height(value) {
+        this.resize(this.width, value);
+    }
+
+    /**
+     * return the width of the canvas which this renderer draws to
+     * @returns {number} width of the system Canvas
+     */
+    get width() {
+        return this.getCanvas().width;
+    }
+
+    set width(value) {
+        this.resize(value, this.height);
+    }
+
+    /**
      * prepare the framebuffer for drawing a new frame
      */
     clear() {}
@@ -127,8 +151,8 @@ export default class Renderer {
         this.cache.clear();
         this.currentScissor[0] = 0;
         this.currentScissor[1] = 0;
-        this.currentScissor[2] = this.getCanvas().width;
-        this.currentScissor[3] = this.getCanvas().height;
+        this.currentScissor[2] = this.width;
+        this.currentScissor[3] = this.height;
         this.clearMask();
     }
 
@@ -139,7 +163,6 @@ export default class Renderer {
     getCanvas() {
         return this.canvas;
     }
-
 
     /**
      * return a reference to this renderer canvas corresponding Context
@@ -194,22 +217,6 @@ export default class Renderer {
     }
 
     /**
-     * return the width of the system Canvas
-     * @returns {number}
-     */
-    getWidth() {
-        return this.getCanvas().width;
-    }
-
-    /**
-     * return the height of the system Canvas
-     * @returns {number} height of the system Canvas
-     */
-    getHeight() {
-        return this.getCanvas().height;
-    }
-
-    /**
      * get the current fill & stroke style color.
      * @returns {Color} current global color
      */
@@ -232,11 +239,10 @@ export default class Renderer {
      */
     overlaps(bounds) {
         return (
-            bounds.left <= this.getWidth() && bounds.right >= 0 &&
-            bounds.top <= this.getHeight() && bounds.bottom >= 0
+            bounds.left <= this.width && bounds.right >= 0 &&
+            bounds.top <= this.height && bounds.bottom >= 0
         );
     }
-
 
     /**
      * resizes the system canvas